server-python/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
requests==2.31.0
pydantic==2.3.0
python-dotenv==1.0.0
openai==1.70.0
